load digital data as int16 (can only be 0 or 1)
I'm not using bool for this

+++ src/spikeanalysis/stimulus_data.py
@@ -303,7 +303,7 @@ class StimulusData:
         except TypeError:
             raise Exception("There is no digital data present")
 
-        values = np.zeros((16, len(self._raw_digital_data)))  # 16 digital-in for intan
+        values = np.zeros((16, len(self._raw_digital_data)), dtype=np.int16)  # 16 digital-in for intan
         for value in range(1, 17):
             values[value - 1, :] = np.not_equal(  # this operation comes from the python Intan code
                 np.bitwise_and(
